use `value_parser`
`allow_invalid_utf8` is deprecated use `value_parser` instead.

The argument for `clap::value_parser!` is an educated guess. I read
about `std::ffi::OsString` from the documentation of `std::env` which I
think is like the type for the command line arguments and it seems to
work but don't quote me on that. This commit should not be done in any
serious project but this is just an exploratory project so I think it's
okay.